refactor: go backend code now has clear packages
There is now a package for database operations, one for logic and one for api operations. The server routes the incoming requests to the appropriate function in the api package, which calls the needed logic to get a result for the response. The logic uses the database to store and get the needed data to get a return value for the api. The database layer just communicates with the database and is more or less a direct abstraction of the sql functions